requirements.txt
google-generativeai
pymongo
streamlit